Recovering all work pushed in the past month.
-overview:
Changes are numerous. This commit is the result of
a cleanup of the git-history via `git reset HEAD~9`
and a forced push to get rid off all merge commits
and messy git-history.

summary of changes:

- update DESCRIPTION file and improve LinkingTo Fields
- provide a vignette that briefly outlines the SV model
equations
- change from snake case to camelCase matching the
coding style of RcppSMC and its accompanying packages
- add better function documenation (for almost all functions)
- improve Particle Gibbs C++ by:
    - adding comments
    - getting rid of unneccessary conversoin between Rcpp and
    arma containers
    - seperate Gibbs sampler part of parameters into separate
    class
    - add options to switch between conditional multinomial,
    stratified, systematic and residual resampling
- improve ancestral line tracking to print only reasonable output
and fixing default parameters for good visual illustration for the
given SV model